attendance_system/requirements.txt
asgiref==3.8.1
Django==5.1.1
djangorestframework==3.15.2
Faker==29.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
six==1.16.0
sqlparse==0.5.1
